+++ asf_search/ASFProduct.py
@@ -1,3 +1,4 @@
+import warnings
 from shapely.geometry import shape, Point, Polygon, mapping
 import json
 
@@ -6,6 +7,9 @@ from asf_search.ASFSearchOptions import ASFSearchOptions
 from asf_search.download import download_url
 from asf_search.CMR import translate_product
 from remotezip import RemoteZip
+
+from asf_search.download.file_download_type import FileDownloadType
+from asf_search import ASF_LOGGER
 
 
 class ASFProduct:
@@ -30,7 +34,7 @@ class ASFProduct:
             'properties': self.properties
         }
 
-    def download(self, path: str, filename: str = None, session: ASFSession = None) -> None:
+    def download(self, path: str, filename: str = None, session: ASFSession = None, fileType = FileDownloadType.DEFAULT_FILE) -> None:
         """
         Downloads this product to the specified path and optional filename.
 
@@ -40,13 +44,45 @@ class ASFProduct:
 
         :return: None
         """
-        if filename is None:
-            filename = self.properties['fileName']
-        
+
+        default_filename = self.properties['fileName']
+
+        if filename is not None:
+            multiple_files = (
+                (fileType == FileDownloadType.ADDITIONAL_FILES and len(self.properties['additionalUrls']) > 1) 
+                or fileType == FileDownloadType.ALL_FILES
+            )
+            if multiple_files:
+                warnings.warn(f"Attempting to download multiple files for product, ignoring user provided filename argument \"{filename}\", using default.")
+            else:
+                default_filename = filename
+                
         if session is None:
             session = self.session
 
-        download_url(url=self.properties['url'], path=path, filename=filename, session=session)
+        urls = []
+
+        def get_additional_urls():
+            output = []
+            base_filename = '.'.join(default_filename.split('.')[:-1])
+            for url in self.properties['additionalUrls']:
+                extension = url.split('.')[-1]
+                urls.append((f"{base_filename}.{extension}", url))
+            
+            return output
+
+        if fileType == FileDownloadType.DEFAULT_FILE:
+            urls.append((default_filename, self.properties['url']))
+        elif fileType == FileDownloadType.ADDITIONAL_FILES:
+            urls.extend(get_additional_urls())
+        elif fileType == FileDownloadType.ALL_FILES:
+            urls.append((default_filename, self.properties['url']))
+            urls.extend(get_additional_urls())
+        else:
+            raise ValueError("Invalid FileDownloadType provided, the valid types are 'DEFAULT_FILE', 'ADDITIONAL_FILES', and 'ALL_FILES'")
+
+        for filename, url in urls:
+            download_url(url=url, path=path, filename=filename, session=session)
 
     def stack(
             self,