+++ asf_search/constants/INTERNAL.py
@@ -1,9 +1,10 @@
 ASF_AUTH_HOST = 'auth.asf.alaska.edu'
 
 CMR_HOST = 'cmr.earthdata.nasa.gov'
-CMR_FORMAT_EXT = 'umm_json_v1_4'
+CMR_FORMAT_EXT = 'umm_json'
 CMR_GRANULE_PATH = f'/search/granules.{CMR_FORMAT_EXT}'
-CMR_COLLECTIONS_PATH = f'/search/collections.{CMR_FORMAT_EXT}'
+CMR_COLLECTIONS = f'/search/collections'
+CMR_COLLECTIONS_PATH = f'{CMR_COLLECTIONS}.{CMR_FORMAT_EXT}'
 CMR_HEALTH_PATH = f'/search/health'
 CMR_PAGE_SIZE = 250
 EDL_HOST = 'urs.earthdata.nasa.gov'
@@ -12,3 +13,5 @@ EDL_CLIENT_ID = 'BO_n7nTIlMljdvU6kRRB3g'
 DEFAULT_PROVIDER = 'ASF'
 
 AUTH_DOMAINS = ['asf.alaska.edu', 'earthdata.nasa.gov']
+
+ERROR_REPORTING_ENDPOINT = 'search-error-report.asf.alaska.edu'