+++ asf_search/download/__init__.py
@@ -1,1 +1,2 @@
 from .download import download_urls, download_url, remotezip
+from .file_download_type import FileDownloadType